requirements.txt
# Core framework
Django==5.2.6

# REST framework
djangorestframework==3.15.0

# JWT authentication
djangorestframework-simplejwt==5.5.1

# Filtering support
django-filter>=24.3

# PostgreSQL driver
psycopg2-binary>=2.9

python-decouple>=3.8

# Testing
pytest>=8.2
pytest-django>=4.8

# Code style / linting
black>=24.4
flake8>=7.1

python-dotenv>=1.0,<2.0   # For .env environment variables

requests>=2.31,<3.0